+++ src/Components/ListResourceComponent.module.css
@@ -1,0 +1,22 @@
+.listContent {
+  background-color: #e4e9f2;
+  border-radius: 10px;
+  border: 1px solid #fff;
+}
+
+.listItem {
+  cursor: pointer;
+}
+
+.listItem:hover {
+  background-color: #f5f5f5;
+
+  border-radius: 10px;
+  border: 1px solid #fff;
+  box-shadow: 0px 2px 8px 0px rgba(0, 0, 0, 0.15);
+}
+
+.pagesButton {
+  display: flex;
+  justify-content: space-between;
+}
