+++ src/Pages/ListPages/EncounterPatientPage.tsx
@@ -1,12 +1,12 @@
-import ObservationService from "../Services/ObservationService";
+import ObservationService from "../../Services/ObservationService";
 import InfoListComponent, {
   InfoListData,
-} from "../Components/InfoListComponent";
+} from "../../Components/InfoListComponent";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 
-import ConditionService from "../Services/ConditionService";
-import PatientQuestionnaireComponent from "../Components/Patient/PatientQuestionnaireComponent";
+import ConditionService from "../../Services/ConditionService";
+import PatientQuestionnaireComponent from "../../Components/Patient/PatientQuestionnaireComponent";
 
 export default function EncounterPatientPage() {
   const observationService = new ObservationService();
